Relax constraint on base package
In order to build this package for M1 Macs, we must use a newer version
of the Stack resolver, but in order to do so, we also have to relax the
version constraint on the base package. Compiling and testing this
package with the latest resolver and base package version worked fine,
so these version constraints should not cause any problems.

We also had to use a newer version of template-haskell to support
lts-20.2